+++ omc3/knob_extractor.py
@@ -61,11 +61,15 @@ The data is fetched from ``NXCALS`` through ``pytimber`` using the **StateTracke
 
 
 """
+from __future__ import annotations
 
 ####### WORKAROUND FOR JAVA ISSUES WITH LHCOP ##################################
 # Move `/mcr/bin` to the end of the Path.
-# This ia a hack, please remove at the earliers convenience. Fixes:
+# Fixes:
 # 'Error: Could not find or load main class (...) aircompressor-0.26.jar'
+#
+# This ia a hack, please remove at the earliers convenience. For updates see: 
+# https://cern.service-now.com/service-portal?id=ticket&table=incident&n=INC3768823
 import os
 if "PATH" in os.environ and "/mcr/bin" in os.environ["PATH"]:
     parts = os.environ["PATH"].split(":")
@@ -80,7 +84,7 @@ import math
 import re
 from datetime import datetime
 from pathlib import Path
-from typing import Dict, List, Optional, Sequence, Tuple, Union
+from typing import TYPE_CHECKING
 
 import pandas as pd
 from dateutil.relativedelta import relativedelta
@@ -91,6 +95,9 @@ from omc3.utils.iotools import PathOrStr, PathOrStrOrDataFrame
 from omc3.utils.logging_tools import get_logger
 from omc3.utils.mock import cern_network_import
 
+if TYPE_CHECKING:
+   from collections.abc import Sequence 
+
 pytimber = cern_network_import("pytimber")
 
 LOGGER = get_logger(__name__)
@@ -100,8 +107,8 @@ ACC_MODELS_LHC = Path("acc-models-lhc")
 KNOBS_FILE_ACC_MODELS = ACC_MODELS_LHC / "operation" / "knobs.txt"
 KNOBS_FILE_AFS = AFS_ACC_MODELS_LHC / "operation" / "knobs.txt"
 
-MINUS_CHARS: Tuple[str, ...] = ("_", "-")
-STATE_VARIABLES: Dict[str, str] = {
+MINUS_CHARS: tuple[str, ...] = ("_", "-")
+STATE_VARIABLES: dict[str, str] = {
     'opticName': 'Optics',
     'beamProcess': 'Beam Process',
     'opticId': 'Optics ID',
@@ -123,7 +130,7 @@ class Head:
     time: str = "EXTRACTION_TIME"
 
 
-KNOB_CATEGORIES: Dict[str, List[str]] = {
+KNOB_CATEGORIES: dict[str, list[str]] = {
     "sep": [
         "LHCBEAM:IP1-SEP-H-MM",
         "LHCBEAM:IP1-SEP-V-MM",
@@ -294,7 +301,7 @@ def main(opt) -> tfs.TfsDataFrame:
 
 # State Extraction -------------------------------------------------------------
 
-def get_state(ldb, time: datetime) -> Dict[str, str]:
+def get_state(ldb, time: datetime) -> dict[str, str]:
     """
     Standalone function to gather and log state data from  the StateTracker.
 
@@ -315,7 +322,7 @@ def get_state(ldb, time: datetime) -> Dict[str, str]:
     return state_dict
 
 
-def _get_state_as_df(state_dict: Dict[str, str], time: datetime) -> tfs.TfsDataFrame:
+def _get_state_as_df(state_dict: dict[str, str], time: datetime) -> tfs.TfsDataFrame:
     """
     Convert extracted StateTracker state-data into a TfsDataFrame
     To be consistent with the main functions output.
@@ -338,7 +345,7 @@ def _get_state_as_df(state_dict: Dict[str, str], time: datetime) -> tfs.TfsDataF
 
 # Knobs Extraction -------------------------------------------------------------
 
-def extract(ldb, knobs: Sequence[str], time: datetime) -> Dict[str, float]:
+def extract(ldb, knobs: Sequence[str], time: datetime) -> dict[str, float]:
     """
     Standalone function to gather data from  the StateTracker.
     Extracts data via pytimber's LoggingDB for the knobs given
@@ -438,7 +445,7 @@ def _extract_and_gather(ldb, knobs_definitions: pd.DataFrame,
     return knobs
 
 
-def _write_knobsfile(output: Union[Path, str], collected_knobs: tfs.TfsDataFrame):
+def _write_knobsfile(output: Path | str, collected_knobs: tfs.TfsDataFrame):
     """ Takes the collected knobs and writes them out into a text-file. """
     collected_knobs = collected_knobs.copy()  # to not modify the df
 
@@ -469,7 +476,7 @@ def _write_knobsfile(output: Union[Path, str], collected_knobs: tfs.TfsDataFrame
 
 # Knobs Definitions ------------------------------------------------------------
 
-def _get_knobs_def_file(user_defined: Optional[Union[Path, str]] = None) -> Path:
+def _get_knobs_def_file(user_defined: Path | str | None = None) -> Path:
     """ Check which knobs-definition file is appropriate to take. """
     if user_defined is not None:
         LOGGER.info(f"Using user knobs-definition file: '{user_defined}")
@@ -487,7 +494,7 @@ def _get_knobs_def_file(user_defined: Optional[Union[Path, str]] = None) -> Path
     raise FileNotFoundError("None of the knobs-definition files are available.")
 
 
-def load_knobs_definitions(file_path: Union[Path, str]) -> pd.DataFrame:
+def load_knobs_definitions(file_path: Path | str) -> pd.DataFrame:
     """ Load the knobs-definition file and convert into a DataFrame.
     Each line in this file should consist of at least three comma separated
     entries in the following order: madx-name, lsa-name, scaling factor.
@@ -537,7 +544,7 @@ def _to_knobs_dataframe(df: pd.DataFrame) -> pd.DataFrame:
     return df
 
 
-def _parse_knobs_defintions(knobs_def_input: Optional[Union[Path, str, pd.DataFrame]]) -> pd.DataFrame:
+def _parse_knobs_defintions(knobs_def_input: Path | str | pd.DataFrame | None) -> pd.DataFrame:
     """ Parse the given knob-definitions either from a csv-file or from a DataFrame. """
     if isinstance(knobs_def_input, pd.DataFrame):
         return _to_knobs_dataframe(knobs_def_input)